+++ src/com/landawn/abacus/android/util/Futures.java
@@ -29,6 +29,7 @@ import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
+import com.landawn.abacus.util.ImmutableIterator;
 import com.landawn.abacus.util.N;
 import com.landawn.abacus.util.Pair;
 import com.landawn.abacus.util.Try;
@@ -532,7 +533,7 @@ public final class Futures {
     }
 
     private static <T> Iterator<T> iterate02(final Collection<? extends CompletableFuture<? extends T>> cfs, final long timeout, final TimeUnit unit) {
-        return new Iterator<T>() {
+        return new ImmutableIterator<T>() {
             private final Iterator<Pair<T, Exception>> iter = iterate22(cfs, timeout, unit);
 
             @Override
@@ -547,11 +548,6 @@ public final class Futures {
                 } catch (InterruptedException | ExecutionException e) {
                     throw N.toRuntimeException(e);
                 }
-            }
-
-            @Override
-            public void remove() {
-                iter.remove();
             }
         };
     }
@@ -590,7 +586,7 @@ public final class Futures {
             });
         }
 
-        return new Iterator<Pair<T, Exception>>() {
+        return new ImmutableIterator<Pair<T, Exception>>() {
             private final int end = cfs.size();
             private int cursor = 0;
 
@@ -612,11 +608,6 @@ public final class Futures {
                 } catch (InterruptedException e) {
                     throw N.toRuntimeException(e);
                 }
-            }
-
-            @Override
-            public void remove() {
-                throw new UnsupportedOperationException();
             }
         };
     }