+++ src/com/landawn/abacus/util/LongMultiset.java
@@ -1108,6 +1108,14 @@ public final class LongMultiset<T> implements Iterable<T> {
         return sortedValues;
     }
 
+    public ImmutableMap<T, Long> toImmutableMap() {
+        return ImmutableMap.of(toMap());
+    }
+
+    public ImmutableMap<T, Long> toImmutableMap(final IntFunction<? extends Map<T, Long>> mapSupplier) {
+        return ImmutableMap.of(toMap(mapSupplier));
+    }
+
     /**
      * 
      * @return a list with all elements, each of them is repeated with the occurrences in this <code>LongMultiset</code>